+++ crates/client/db/src/meta_db.rs
@@ -9,7 +9,7 @@ use sp_runtime::traits::Block as BlockT;
 
 use crate::DbHash;
 
-#[derive(Encode, Decode)]
+#[derive(Debug, Clone, Encode, Decode)]
 pub struct L1L2BlockMapping {
     pub l1_block_hash: H256,
     pub l1_block_number: u64,
