+++ crates/client/state-sync/src/tests/mod.rs
@@ -1,5 +1,4 @@
 //! Madara client testing utilities.
 pub mod constants;
 pub mod helpers;
-pub mod l1;
 pub mod sync;
